requirements.txt
altair
appdirs
attrs
beautifulsoup4
bcrypt
blinker
cachetools
certifi
chardet
charset-normalizer
click
colorama
contourpy
cycler
decorator
docopt
entrypoints
extra-streamlit-components
Faker
favicon
fonttools
frozendict
gitdb
GitPython
htbuilder
html5lib
idna
importlib-metadata
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
matplotlib
mdurl
more-itertools
multitasking
numpy
oauthlib
packaging
pandas
peewee
Pillow
pipreqs
plotly
protobuf
pyarrow
pydeck
Pygments
PyJWT
Pympler
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
PyYAML
referencing
requests
requests-file
requests-oauthlib
rich
rpds-py
scraper
selenium
six
smmap
soupsieve
st-annotated-text
streamlit
streamlit-authenticator
streamlit-camera-input-live
streamlit-card
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-toggle-switch
streamlit-vertical-slider
tenacity
threadpoolctl
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
xlrd
yarg
yfinance
zipp
huggingface_hub
python-dotenv
gtts
speechrecognition
pypdf2
langchain
langchain_google_genai
google-generativeai
faiss-cpu
scikit-learn
xgboost
ta
psycopg2-binary
fpdf
streamlit_folium
numerize
web3
ccxt
langchain_google_genai
PyPDF2
gtts
langchain==0.3.1
pydantic==2.7.4
langchain_community
statsmodels
seaborn
streamlit_option_menu
openpyxl
